scripts/kafka/requirements.txt
kafka-python
python-dotenv